+++ CHANGELOG.md
@@ -8,6 +8,16 @@
 - [1.0.1](#101---2021-05-12)
 - [1.0.0](#100---2021-05-01)
 
+## [Unreleased][]
+### Added
+- Biography Integration: now works with Biograpy 1.2.X version too. 'Notify tags' must be enabled on Bio's config panel, the script will do it automatically when installing the mod. Appart from that, installation and selection mode sync works the same than 1.1.X series.
+### Changed
+- Portable: first time properties are set use now relative paths on profile folder for portable installations (>= 1.6).
+- Data: json file for tags is now formatted to be readable.
+- Biography Integration: manually reloading Biography panel after installing the mod is no longer required, since it's done auttomatically after installing the mod.
+### Removed
+### Fixed
+
 ## [1.2.0] - 2021-06-07
 ### Added
 ### Changed